feat: реализовать валидацию и обработку исключений

+++ src/main/java/com/batin/wallet/Wallet.java
@@ -1,6 +1,7 @@
 package com.batin.wallet;
 
 import jakarta.persistence.*;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -19,5 +20,5 @@ public class Wallet {
     @Id
     @GeneratedValue
     private UUID id;
-    private int balance;
+    private long balance;
 }
